[#884] Code cleanup and iox2dump and hex record & replay
